Two pointer Approach the crux of two pointer approach is whenever the answer you are looking to find relies on two different parameters of lets say an array/any other data structure, using this technique may result in optimal solution.
Like in this case since we need to find maximum size of container
starting at two opposite ends makes sense because largest container will
obviously have largest width.

Similarly in case of reversing an array/ finding palindrome. It is
important look for keywords such as largest, which can help you to
decide if this technique is appropriate.

Overall great problem, learned a lot and many more to come.